[ENSCORESW-594]. Allow regulatory regions to be queried from an ID.
The regulatory REST code relies on a species being set in the stash. This
is not the case when we query with an ID. Code has been changed to add
species to the stash. We ran this against live and it works. Normal
testing is not possible as we do not have a regulation DB in our
test suite yet.