[ENSCORESW-535]. Fixed lookup issues as seen in ensembl genomes.
A number of display xrefs in ensembl genomes are not linked via
object_xref to the object's hierachy. This is for a number of
complex reasons. If we do not find a symbol then we should switch
to looking for the display_label linked to a gene. Do not do the
search if we did not ask for a gene.