[ENSCORESW-643]. Supporting GFF3 from the translation features endpoint.
This was completed but fell down when new data was added to the endpoint. This
now fakes a Slice, lets a Translation act as a region to annotate to and brings
in new model objects to handle the translation into GFF3. We also have some
somatic variation changes coming along for the ride because it is impossible
to disentangle this.