[ENSCORESW-656]. Fixing the test suite before release.
We found a number of issues with the test suite. Seems that the
tests were not working before I handed this code over. We now
use the info endpoint for compara info tests. This meant the
removal of the compara.t tests since they only checked the info
nothing else.

Missing imports were found in ExternalDB.

Documentation was updated as was the default values
for the test site.