[ENSCORESW-628]. Supporting a JSON output for GeneTrees.
Finally we have a JSON output for gene trees. it's not done nicely but it works. Also it is very very fast! Support is done using the TreeHash builder and then delegating to the JSON encoder. A custom serialiser has been built in the view layer. Otherwise we would have to do content-type/accepts detection. That's not nice at all!