[ENSCORESW-774]. Better default compara now chosen.
We had an error whenever we used the ensembl genomes species with the
multi databases from compara. Our best guess compara DB was switching
to the division’s compara DB rather than using multi. We now avoid
that problem by giving a default. So we will still try to use the
division’s DB but the lack of it will result in the correct results
being passed back.