[ENSCORESW-847] & [ENSCORESW-848]. Improved BED support for transcripts.
Issue one was the thick and thin ends did not pay any attention to strandedness
so the values extracted and reported were wrong. This error was picked up when
trying to create BigBEDs from the REST output.

Second issue was the speed to emit this transcript data was so slow. Turns
out we never forced the pre-loading of transcripts in bed format. Turning this
option on in the fetch code brought the runtime down from ~9-12 seconds to
~2-3 seconds (x4 speedup). The load exons only triggers when we are in BED
and using transcripts.