Fail to build a rate limiter if there is no backend.
Earlier versions of this assumed you could default fail over to a
simplebackend. Production showed this is not the case. There is no
reason for a failover and was a bad idea on my part. Now we croak if we
see there is not one given. Test suites have been updated accordingly.