JSONP was specified as a output format causing documentation problems.
We specified jsonp as an output format. This caused problems in the
documentation code which looked to see if JSONP was already specified as
an output format. It meant that callback was not added as a parameter
which is a bit poor. Removing jsonp as an output format fixes the error.