Backup global distance variables
UPSTREAM_DISTANCE and DOWNSTREAM_DISTANCE are used by VEP to
determine the max distance between a VF and a transcript to
call a consequence type. This can be altered by a user in a
plugin, but unless we back up the original value the modified
value will persist until the server is restarted, thereby
affecting future calls.