fix: force lookup on correct object type
In Genomes there are objects with different types but the same stable id.
When we try to search for such object by stable id the first record is
picked. That means sometimes we have an object with correct stable id but
wrong type.

This is also affecting Ensembl too:
http://rest.ensembl.org/map/translation/F53H8.4/100..300?content-type=application/json

Make sure that we lookup records with correct object type accroding to endpoints.

Fixes ENSEMBL-4667.