Add searching compara dbs for a gene id at homology endpoint
In EnsemblGenomes we have vertebrate gene ids stored in compara pan homology
database. These ids are not loaded into stable_id_lookup database. When a user
wants to fetch homologues through a vertebrate gene id, it returns error because
the id is not found in stable_id_lookup database.

This change allows searching compara dbs at homology endpoint so that when a
vertebrate gene id is not found in the stable_id_lookup database, we continue to
search for it in compara dbs, so that the id could be used to fetch homologues
through EnsemblGenomes's rest endpoint.

Related to ENSEMBL-5061.