cpanfile: Require YAML::XS, YAML::Syck and URI::Find
These three modules are required for the YAML (the former) and HTML (the
latter two) serialisers provided by Catalyst::Controller::REST to work.
The latter is especially important because it is our fallback
serialiser for unknown content types, which in our case includes e.g.
text/x-fasta; its presence is therefore necessary for erroneous requests
for our special content types (such as requesting a sequence for an
unknown transcript stable ID) to return the actual error message, e.g.

	{"error":"ID 'ENSTdeadbeef' not found"}

(the content type of error responses to requests involving content types
handled by the fallback serialiser is 'application/json') instead of the
generic (and pretty much useless)

	Content-Type text/html is not supported.