Registry: warn instead of dying if creating an EQTLAdaptor has failed
This improves two things.

One, which is what this is for in the first place: if the server is
configured to both load eQTL files and preload the Registry (and it is
worth pointing out that documentation recommends enabling the latter if
eQTL endpoints are to be enabled) but the eQTL file is not available
when the server is started, in the past the whole server would fail to
start; see ENSCORESW-2670. This is because unlike the endpoint code,
the preloader did attempt to capture exceptions originating from
_get_eqtl_adaptor(). Now the server does start even if preloading

Two, which is why it has been added inside _get_eqtl_adaptor() rather
than in the preloader: with preloading disabled, querying a
non-functioning eQTL endpoint would sometimes forward the internal error
message (such as "No such file or directory at /path/to/EQTLAdaptor.pm
line 266") from EQTLAdaptor::new() to the user. Even leaving aside the
security implications of exposing the server's internals to the
Internet, such information is at best useless and at worst confusing to
users. Now that the error is caught and its message rebroadcast as a
warning it is only printed on the server side (e.g. on the console),
whereas the users get standard "No EQTL adaptor available" REST error.