Model::Lookup::find_object_location(): only pass 3 args to sprintf()
The sprintf() call in line 350 of Lookup.pm expects three arguments to
insert into the string yet the @captures array it retrieves them from is
longer than that, resulting in a "redundant arguments" warning. This
warning appears A LOT in the test-suite output. Slice @captures to only
pass 3 substitution arguments to sprintf(), and just to be on the safe
side change the preceding test for the truthiness of its first element
to that of the third one.